requirements.txt
Telethon
pyautogui
requests
python-dotenv